Fix blank pages in ViewtypeEditor
Previous behavior after loading an EViewtype file: the Source (text editor) page
is open, and the other pages are blank.  Unless, mysteriously, I open an XMI
file.  This seems to trigger something, and all pages are visible.

Trying to isolate the problem, I found that using ScrolledForm into the
MultiEditorPart was the issue.

Converting from a MultiEditorPart to a FormEditor did the trick.  This seems to
be the standard way of doing a PDE-like multi-page editor.