Get filters from the virtual links XMI
Previously, we were pulling the elements to be filtered from an Ecore file.

Now, we use the Filter class from the virtual links metamodel.

I've used the elementRef of the linkedElement in a way that's maybe not
orthodox.  Maybe we could use URIs or something, but it works well for the
examples.