Move test resources to the test plugin; refactor URI loading
Moving the resources to the test plugin simplifies running the tests.
Previously, we had a to create a dummy test workspace containing the a project
with the resources.  Now, the resources are loaded directly from a folder in the
test plugin with relative URI.

To make this work, we needed to change the `createPlatformResourceURI` calls to
`createURI`.  This also makes the path in the eview and eviewpoint files more
flexible, as we can still use platform URI with the `platform:` prefix, but now
they support relative paths for the tests.

Also, this commits renames the `full` test to `three model composition`, and
reorganizes the test resources folder.