Add virtual references to specific subtypes of VirtualLink
Since that's basically what we did previously: separate the concrete types into
different lists.  Now that's part of the VirtualLink API.

The implementation follows the strategy suggested in the EMF bible (Feature
Maps, §8.1.1).  These virtual references are immutable, so you have to use the
virtualLinks containment list to add new concrete links.  This has the advantage
of enforcing a definite order on the elements.