Specify order of packages in the viewpoint
Previously the order was non-deterministic, and based on the order of the
underlying hashtable used by the package registry of the virtual resource set.

Now, the order of packages in the contents of the viewpoint resource is the
order of the contributing models in the .eviewpoint file.  If there are any new
concepts in the weaving model, they are put in a new virtual package that
appears at the end.