Support adding new concepts
All new concepts go in the same virtual package.  That package's name is taken
from a new name attribute on the WeavingModel class.

This commit adds the name attribute in VirtualLinks.ecore, and updates the
generated code.  As the name is mandatory attribute, the models used by the
tests are updated as well.

In Viewpoint, we create the package if the name is valid and if there are any
new concepts.  Then we add the new concepts to it.