Better error message when the Ecore editor fails to load a Viewpoint
Since the default ResourceImpl.load implementation wraps the doLoad in `try` and
silently disposes of exception, we have no way to signal to the user that
something went wrong with the Viewpoint configuration and we could not build the
contents.

So, instead of returning an empty contents list, or a list with the contributing
metamodels intact, we return a custom exception.