Remove package from path of elements in Viewpoint
The reasoning is that the package name is not necessary.  In the weaving model,
paths are in a concrete element, that is already part of a model, so we already
know the package name.  When dealing with EMF directly, we usually have an
EPackage to give as root to findElement, so there's no need to check this
information in the path.