Load Ecore metamodels from Viewpoint in the global registry
This is a better solution to the problem solved by the previous commit.

ECL has to find the metamodels.  If we give ECL the path to an Ecore file, it
will have to load the file all other again, which is work that Viewpoint already
did.

It seems just simpler to have the Viewpoint load an Ecore into the global
registry once and for all.