Make sure to close contributing resources on unload
NeoEMF resources need to be properly closed, since they have a lock file
that need to be released.

We can do this easily enough in EView Editor.  But unfortunately other
editors that could deal with views (Sample Ecore Editor/Modisco) do not
call `resource.unload`, so we have to be careful not to use them until
there is a workaround.