Remove idempotence of virtualizer
It was the wrong solution to my uncarelessness.  We should rather
construct the viewpoint to ensure that getVirtual is called only when
necessary.

Also, the function was not idempotent, since it prevented the creation
of /all/ virtual objects regardless of whether they created by
getVirtual call to the same virtualizer.

Removing idempotence should allow for viewpoints on viewpoints, at any
level.