Bump a few dependencies for the maven build
The goal is to make sure we are building correctly for the Eclipse 4.9
release.

Bumping to Xtext 2.15 also lets us get rid of the workaround of pinning
equinox for running MWE2.

Unfortunately, due to an ongoing bug [1], we have to make all our tests
bundles depend explicitly on javax.annotation.

[1]: https://bugs.eclipse.org/bugs/show_bug.cgi?id=539038