Require alias for contributing(Meta)models in eviewpoint and eview
Switching to InMemoryEmfView for the ECL delegate created an issue where
classifiers with the same name would be resolved to the first metamodel
in the resource set.  However, we also loaded all metamodels in the
resource set for each resource, since we had no way to know which
metamodels were used for which models based on the information in the
eviewpoint/eview file.  This made the second rule in the traceability
demo silently fail, because the `Package` class is present in both UML
and Java.

The fix is to be explicit about the metamodel-model link in the
eviewpoint and eview files.  We know require a prefix alias in the
contributing(Meta)models lines.

The bonus side-effect is that we don't need to put these aliases in the
ECL header comments anymore, so we can get rid of the ad-hoc parsing
there.