Lazily match ECL rules
The O(n^2) algorithm for matching ECL rules was never a good thing, even
for small models.

Now, ECL files are parsed when the view is loaded, but rules are not
executed until the corresponding feature is accessed for the first
time.  Then the rule is executed, but only for the object owning that
feature, so we have O(n) matching, and the result is cached.